services/requirements.txt
paho-mqtt
psycopg2-binary
schedule